Handle org....BookView.setFlags api in EDataBookView.
This handles the dbus api and just sets up the flags on
the private data. Currently the patch assumes that the
flags will be set before starting the view however it
would be good eventually to add a "flags" property to
the EDataBookView so that backends could watch it for
change notifications.