Added test case showing suppression of initial notifications.
Added the 'test-ebook-suppress-initial-notifications' test, this test
essentially opens an EBookView and sets the flags to 0 (it unflags the
default E_BOOK_VIEW_NOTIFY_INITIAL bit) so that the initial content is
not sent in notifications.

The test asserts that no notifications are sent before 'completed' is
signaled and then add's a contact and asserts that the view does
receive notification for the newly added contact.