Fix "avoid initial notifications" functionality
The file backend does need to notify the view so that it can
update its hash table of ids. The filtering needs to be done in
EDataBookView instead, AFTER the hash table is updated.

This way, the functionality does not require any change in the
backend and works for all backends.

Bug report:
https://bugzilla.gnome.org/show_bug.cgi?id=652171