Fix stray unref of server in imapx add_folders_to_summary()
This was causing a double (well, multiple) free and use-after-free of the
server; it has no business here.

It was actually seen when a broken Yahoo server gave a namespace with NIL
for the dir_sep, thus causing problems with subfolders.
(cherry picked from commit c3460e79201ba988500014386dbc3f8781dbc5f3)