CamelIMAPXServer: Add new namespace and mailbox APIs.
CamelIMAPXMailbox and CamelIMAPXNamespace are closely intertwined, so
the containing data structures for them need to be introduced together.

This commit does so, and populates the data structures in memory from
untagged server responses.  It also adds CamelIMAPXServer API to access
or query the data structures.

All of this is so far unutilized.  Actually integrating these APIs into
the other parts of IMAPX will take place in subsequent commits.

New functions:

  camel_imapx_server_ref_namespaces()
  camel_imapx_server_ref_mailbox()
  camel_imapx_server_ref_selected()
  camel_imapx_server_list_mailboxes()

Removed functions:

  imapx_parse_namespace_list()