CamelIMAPXServer: Support the SPECIAL-USE capability.
http://tools.ietf.org/html/rfc6154

Camel does not yet provide an API for applications to utilize this
information, but we can still capture it in a CamelIMAPXMailbox.

I'm still pondering a suitable application-facing API for this.  I'd
prefer to avoid stuffing this into CamelFolderInfoFlags if I can, as
Camel already overuses / abuses flag values as it is.